Version 1.2.10
02.09.2001 - Added function ftp_keepalive(). This is used to keep the
             control connection open on very slow connections that
             transmit large files. Some firewalls close the control
             connection if there is no activity for a longer time.
10.09.2001 - Function get_remote_file_names() sorted the structure
             retrieve_list wrongly. This could cause gf_ftp to crash. Fixed.
11.09.2001 - Hostname was not a compound string in afd_info dialog.
             the date of the files so that when the age-limit option
             is set files will not be deleted by sf_xxx.
17.09.2001 - New parameter (CREATE_SOURCE_DIR_MODE) in AFD_CONFIG file to
             set the mode of the input directories created by AFD.
22.09.2001 - Catch fifo buffer overflows in transfer_log and receive_log.
29.09.2001 - Added one more row to choose from in afd_ctrl and mon_ctrl
             dialog (from Setup menu).
	     we do have the correct permission to update the file. If
	     not copy the file.
             with the trans_rename option would not be resend. Wonder
             why no one has noticed this!
30.09.2001 - Updated SCP1 functions so renaming works.
01.10.2001 - Dialog view_dc and show_olog did not show the time option
             if specified for that directory. Fixed.
09.10.2001 - Added option to add hostname for rename and trans_rename
             option.
12.10.2001 - Did not reduce priority of a job if the remote directory
             is no longer available. Fixed.
15.10.2001 - Fixed some minor bugs in mon_ctrl dialog.
18.10.2001 - Process FD failed to reset FSA values when a job is removed
             because it is no longer in the DIR_CONFIG.
19.10.2001 - Compiles under HP 10.20 again.
23.10.2001 - Fixed a rename bug in function change_name().
26.10.2001 - Remove rotating dash in job_details dialog, when file has
             been completely transmitted.
28.10.2001 - Compatibility with lesstif has been improved. However there
             are still some issues remaining: drag+drop in edit_hc
             does not work and displaying logs with show_log is still
             problematic.

+++ fd/sf_loc.c
@@ -411,7 +411,7 @@ main(int argc, char *argv[])
             {
                if (errno == EEXIST)
                {
-                  if (unlink(p_to_name) == -1)
+                  if ((unlink(p_to_name) == -1) && (errno != ENOENT))
                   {
                      trans_log(ERROR_SIGN, __FILE__, __LINE__,
                                "Failed to unlink() <%s> : %s",
@@ -424,9 +424,12 @@ main(int argc, char *argv[])
                   }
                   else
                   {
-                     trans_log(INFO_SIGN, __FILE__, __LINE__,
-                               "File <%s> did already exist, removed it and linked again.",
-                               p_to_name);
+                     if (errno != ENOENT)
+                     {
+                        trans_log(INFO_SIGN, __FILE__, __LINE__,
+                                  "File <%s> did already exist, removed it and linked again.",
+                                  p_to_name);
+                     }
 
                      if (link(source_file, p_to_name) == -1)
                      {
